requirements-test.txt
flake8
coverage
pyquery
factory-boy
